ignore(run_config_test.py): 添加新忽略文件并优化日志目录删除逻辑
- 在 .gitignore 文件中添加 run_config_test.py，避免跟踪该测试脚本
- 修改 out_clean.py，增加对 log目录存在性的检查再进行删除
- 新增 worker_time.py脚本，用于测试不同参数下数据加载器的性能

+++ out_clean.py
@@ -31,4 +31,5 @@ if __name__ == "__main__":
     delete_unmatched_folders(checkpoints_dir, results_dir)
     delete_unmatched_folders(test_results_dir, results_dir)
     log_dir = Path("./out/log")
-    shutil.rmtree(log_dir)
+    if log_dir.exists():
+        shutil.rmtree(log_dir)
